initial commit
This is a driver / service for the bcm2835 mailbox which can be found on
certain Raspberry Pi models as interface to the firmware. The service
implements the IO server vbus protocol allowing to share access to the
bcm2835 between clients.

The client interface is very similar to direct hardware access so that
it's possible to implement direct hardware access and access using this
service can be done with low effort.

Client filtering is done using white lists written in Lua.

Change-Id: I73e14d462a7383d74e2420bb8e5420502affc7b5